+++ src/agents/rl/qlearning/QLearner.hpp
@@ -17,17 +17,17 @@
 
 class QLearner : public RLLearner{
 	private:
-		double alpha, delta, lambda, traceThreshold;
+		float alpha, delta, lambda, traceThreshold;
 		int numFeatures, currentAction, nextAction;
 
 		std::string nameWeightsFile;
 
 		vector<int> F;					//Set of features active
 		vector<int> Fnext;              //Set of features active in next state
-		vector<double> Q;               //Q(a) entries
-		vector<double> Qnext;           //Q(a) entries for next action
-		vector<vector<double> > e;      //Eligibility trace
-		vector<vector<double> > w;      //Theta, weights vector
+		vector<float> Q;               //Q(a) entries
+		vector<float> Qnext;           //Q(a) entries for next action
+		vector<vector<float> > e;      //Eligibility trace
+		vector<vector<float> > w;      //Theta, weights vector
 		vector<vector<int> >nonZeroElig;//To optimize the implementation
 		
 		/**
@@ -45,10 +45,10 @@ class QLearner : public RLLearner{
 		/**
  		* In Q-Learning the Q-values (one per action) are updated as the sum of weights for that given action.
  		* To avoid writing it more than once on the code, its update was extracted to a separate function.
- 		* It updates the vector<double> Q assuming that vector<int> F is filled, as it sums just the weights
+ 		* It updates the vector<float> Q assuming that vector<int> F is filled, as it sums just the weights
  		* that are active in F.
  		*/
-		void updateQValues(vector<int> &Features, vector<double> &QValues);
+		void updateQValues(vector<int> &Features, vector<float> &QValues);
 		
 		/**
  		* When using Replacing traces, all values not related to the current action are set to 0, while the
