+++ src/agents/rl/true_online_sarsa/TrueOnlineSarsaLearner.cpp
@@ -30,8 +30,8 @@ TrueOnlineSarsaLearner::TrueOnlineSarsaLearner(ALEInterface& ale, Features *feat
 		Q.push_back(0);
 		Qnext.push_back(0);
 		//Initialize e:
-		e.push_back(vector<double>(numFeatures, 0.0));
-		w.push_back(vector<double>(numFeatures, 0.0));
+		e.push_back(vector<float>(numFeatures, 0.0));
+		w.push_back(vector<float>(numFeatures, 0.0));
 
 		nonZeroElig.push_back(vector<int>());
 	}
@@ -43,9 +43,9 @@ TrueOnlineSarsaLearner::TrueOnlineSarsaLearner(ALEInterface& ale, Features *feat
 
 TrueOnlineSarsaLearner::~TrueOnlineSarsaLearner(){}
 
-void TrueOnlineSarsaLearner::updateQValues(vector<int> &Features, vector<double> &QValues){
+void TrueOnlineSarsaLearner::updateQValues(vector<int> &Features, vector<float> &QValues){
 	for(int a = 0; a < numActions; a++){
-		double sumW = 0;
+		float sumW = 0;
 		for(unsigned int i = 0; i < Features.size(); i++){
 			sumW += w[a][Features[i]];
 		}
@@ -53,7 +53,7 @@ void TrueOnlineSarsaLearner::updateQValues(vector<int> &Features, vector<double>
 	}
 }
 
-void TrueOnlineSarsaLearner::updateWeights(int action, double alpha, double delta_q){
+void TrueOnlineSarsaLearner::updateWeights(int action, float alpha, float delta_q){
 	for(unsigned int a = 0; a < nonZeroElig.size(); a++){
 		for(unsigned int i = 0; i < nonZeroElig[a].size(); i++){
 			int idx = nonZeroElig[a][i];
@@ -67,8 +67,8 @@ void TrueOnlineSarsaLearner::updateWeights(int action, double alpha, double delt
 	}
 }
 
-void TrueOnlineSarsaLearner::updateTrace(int action, double alpha){
-	double dot_e_phi = 0;
+void TrueOnlineSarsaLearner::updateTrace(int action, float alpha){
+	float dot_e_phi = 0;
 	for(unsigned int i = 0; i < F.size(); i++){
 		int idx = F[i];
 		dot_e_phi += e[action][idx];
@@ -147,11 +147,11 @@ void TrueOnlineSarsaLearner::loadWeights(){
 void TrueOnlineSarsaLearner::learnPolicy(ALEInterface& ale, Features *features){
 	
 	struct timeval tvBegin, tvEnd, tvDiff;
-	vector<double> reward;
-	double elapsedTime;
-	double norm_a;
-	double q_old, delta_q;
-	double cumReward = 0, prevCumReward = 0;
+	vector<float> reward;
+	float elapsedTime;
+	float norm_a;
+	float q_old, delta_q;
+	float cumReward = 0, prevCumReward = 0;
 	unsigned int maxFeatVectorNorm = 1;
 	sawFirstReward = 0; firstReward = 1.0;
 
@@ -227,11 +227,11 @@ void TrueOnlineSarsaLearner::learnPolicy(ALEInterface& ale, Features *features){
 		}
 		gettimeofday(&tvEnd, NULL);
 		timeval_subtract(&tvDiff, &tvEnd, &tvBegin);
-		elapsedTime = double(tvDiff.tv_sec) + double(tvDiff.tv_usec)/1000000.0;
+		elapsedTime = float(tvDiff.tv_sec) + float(tvDiff.tv_usec)/1000000.0;
 		
-		double fps = double(ale.getEpisodeFrameNumber())/elapsedTime;
+		float fps = float(ale.getEpisodeFrameNumber())/elapsedTime;
 		printf("episode: %d,\t%.0f points,\tavg. return: %.1f,\t%d frames,\t%.0f fps\n",
-			episode + 1, cumReward - prevCumReward, (double)cumReward/(episode + 1.0),
+			episode + 1, cumReward - prevCumReward, (float)cumReward/(episode + 1.0),
 			ale.getEpisodeFrameNumber(), fps);
 		totalNumberFrames += ale.getEpisodeFrameNumber();
 		prevCumReward = cumReward;
@@ -240,9 +240,9 @@ void TrueOnlineSarsaLearner::learnPolicy(ALEInterface& ale, Features *features){
 }
 
 void TrueOnlineSarsaLearner::evaluatePolicy(ALEInterface& ale, Features *features){
-	double reward = 0;
-	double cumReward = 0; 
-	double prevCumReward = 0;
+	float reward = 0;
+	float cumReward = 0; 
+	float prevCumReward = 0;
 
 	//Repeat (for each episode):
 	for(int episode = 0; episode < numEpisodesEval; episode++){
@@ -260,7 +260,7 @@ void TrueOnlineSarsaLearner::evaluatePolicy(ALEInterface& ale, Features *feature
 		ale.reset_game();
 		sanityCheck();
 		
-		printf("%d, %f, %f \n", episode + 1, (double)cumReward/(episode + 1.0), cumReward-prevCumReward);
+		printf("%d, %f, %f \n", episode + 1, (float)cumReward/(episode + 1.0), cumReward-prevCumReward);
 		
 		prevCumReward = cumReward;
 	}
